The settings routes live in app/api/settings.py; the tests that fake them patch that module, proven by a canary
