+++ app/api/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI routers split out of app/main.py, one module per screen area."""
